Move getFormattedEarnedReward to a util
